added phase reference finder
phase reference finder obtains a reference from the channel estimation and not from the coarse frequency estimator as in the case of the ofdm_frame_equalizer_adapt, which becomes useless with this replacement;
Although it seems to behave in a proper way, I think it would be important to add tags to help debug the block. 
The block also seems to have problems with the proportion of inputs to outputs, this has to be double checked!